Fix vehicle association not associating correctly
I haven't read the story description but, from what I can see, there is
no check that compare the currently associated driver of a vehicle to
the driver passed as a parameter. It might be a good idea to do this
check to not remove an association of two unrelated entities.